+++ deploy-platform/src/lib/workflows/Github.types.ts
@@ -29,8 +29,6 @@ export interface DispatchOptions {
   ref: string;
   /** Id */
   stepUUID: StepUUID;
-  /** Environment */
-  environment: string;
 }
 
 /** Options for the dispatch method */